gauntlet: Fix melee punch/scepter accuracy
accLance, accWhip, and accTent are all being used as placeholders for
gauntlet melee specific things (punch/scepter). However, the accuracy
values they have reflect their actual name and not their intended
function. This commit fixes the values to be correct by removing the
overload boost (120 -> 99) on all 3 and the multiplier bonus on
accLance/naked punch (1.1 -> 1).

To minimize noise, the names are kept the same. It would be useful to
change these to something more meaningful in a follow up commit.