gauntlet: Fix player magic/ranged def rolls
While praying Augury, the player defensive roll against magic is higher
than the magic pdr when praying Rigour/Piety. This initial logic was
encoded, but the logic for switching between def rolls was missing and
has now been added.

For calculating the adjusted def level for the magic def rolls, the
levels need to be individually computed and floored, before being
combined. At lvl 99 def/magic with Piety/Rigour, the order does not make
a difference and the adjusted level is 106. At lvl 98 though, it does
make a difference (104 vs 105). So that change is not strictly
necessary, but may be useful in case this logic is reused elsewhere.

Nit: My understanding is that the Augury def roll against magic should
also be slightly higher, as the gauntlet staff provides a style bonus of
+3 to magic. I left this out for simplicity as it would require tracking
and switching between additional def rolls.